lestarch: renaming ambiguous files
